Cow breed model and refactors.
